merge

+++ kbe/src/server/dbmgr/dbtasks.cpp
@@ -337,7 +337,7 @@ bool DBTaskWriteEntity::db_thread_process()
 
 	if(writeEntityLog)
 	{
-		if (pDatas_->length() < (sizeof(uint32) + sizeof(uint32)))
+		if (pDatas_->length() < (sizeof(uint32) + sizeof(uint16)))
 		{
 			ERROR_MSG(fmt::format("DBTaskWriteEntity::db_thread_process(): MemoryStream exception(rpos={}, wpos={})! entityID={}, sid={}, callbackID={}, shouldAutoLoad={}, address={}\n",
 				pDatas_->rpos(), pDatas_->wpos(), eid_, sid_, callbackID_, shouldAutoLoad_, addr_.c_str()));
